+++ src/com/intellij/idea/plugin/hybris/impex/lang/annotation/ImpexAnnotator.kt
@@ -27,11 +27,12 @@ import com.intellij.idea.plugin.hybris.impex.psi.references.ImpExHeaderAbbreviat
 import com.intellij.idea.plugin.hybris.impex.psi.references.ImpexMacroReference
 import com.intellij.idea.plugin.hybris.lang.annotation.AbstractAnnotator
 import com.intellij.lang.annotation.AnnotationHolder
-import com.intellij.lang.annotation.HighlightSeverity
+import com.intellij.openapi.editor.markup.HighlighterLayer
 import com.intellij.openapi.util.TextRange
 import com.intellij.psi.PsiElement
 import com.intellij.psi.util.elementType
 import com.intellij.psi.util.parentOfType
+import org.jetbrains.kotlin.idea.codeinsight.utils.findExistingEditor
 
 class ImpexAnnotator : AbstractAnnotator(DefaultImpexSyntaxHighlighter.instance) {
 
@@ -54,15 +55,21 @@ class ImpexAnnotator : AbstractAnnotator(DefaultImpexSyntaxHighlighter.instance)
     override fun annotate(element: PsiElement, holder: AnnotationHolder) {
         when (element.elementType) {
             ImpexTypes.VALUE_LINE -> {
-                val valueLine = element as? ImpexValueLine ?: return
-                val headerLine = valueLine.headerLine ?: return
-                val valueLineIndex = headerLine.valueLines.indexOf(valueLine)
+                element.findExistingEditor()
+                    ?.let {
+                        val valueLine = element as? ImpexValueLine ?: return
+                        val valueLineIndex = valueLine.headerLine
+                            ?.valueLines
+                            ?.indexOf(valueLine)
+                            ?: return
+                        val lineNumber = it.document.getLineNumber(valueLine.textOffset)
 
-                if ((valueLineIndex + 1) % 2 == 0) {
-                    highlight(ImpexHighlighterColors.VALUE_LINE_EVEN, holder, element, HighlightSeverity.INFORMATION)
-                } else {
-                    highlight(ImpexHighlighterColors.VALUE_LINE_ODD, holder, element, HighlightSeverity.INFORMATION)
-                }
+                        if ((valueLineIndex + 1) % 2 == 0) {
+                            it.markupModel.addLineHighlighter(ImpexHighlighterColors.VALUE_LINE_EVEN, lineNumber, HighlighterLayer.SYNTAX)
+                        } else {
+                            it.markupModel.addLineHighlighter(ImpexHighlighterColors.VALUE_LINE_ODD, lineNumber, HighlighterLayer.SYNTAX)
+                        }
+                    }
             }
 
             ImpexTypes.USER_RIGHTS_HEADER_PARAMETER -> {