+++ src/com/intellij/idea/plugin/hybris/system/cockpitng/lang/folding/CngConfigFoldingBuilder.kt
@@ -88,7 +88,7 @@ class CngConfigFoldingBuilder : AbstractXmlFoldingBuilderEx<CngFoldingSettings, 
                     ?.value
                     ?.trimmedText
                     ?.let { " = $it"}
-                    ?: "?"
+                    ?: " ?"
 
                 name + value
             }