+++ src/com/intellij/idea/plugin/hybris/tools/remote/http/HybrisHacHttpClient.java
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -54,7 +55,7 @@ public class HybrisHacHttpClient extends AbstractHybrisHacHttpClient {
     private static final Logger LOG = Logger.getInstance(HybrisHacHttpClient.class);
 
     public static HybrisHacHttpClient getInstance(@NotNull final Project project) {
-        return project.getComponent(HybrisHacHttpClient.class);
+        return project.getService(HybrisHacHttpClient.class);
     }
 
     public @NotNull