+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added inline collection merge modes: `(+)`, `(-)`, `(+?)` [#671](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/671)
 - Added colorization for `(+)`, `(-)` and `(+?)` inline collection merge modes [#672](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/672)
 - Added colorization for unique header parameter name [#683](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/683)
+- Improved colorization for `odd` and `even` value lines [#710](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/710)
 - Added support of the `<null>` special value [#673](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/673)
 - Automatically add `=` after type and attribute modifier and trigger code completion [#674](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/674)
 - Automatically trigger code completion after choosing the mode [#675](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/675)